server/srealitySpider/requirements.txt
psycopg2-binary==2.9.6
Scrapy==2.9.0
